Fix : myPageController
mypage 컨트롤러에 mypage 관련 컨트롤러를 모아 정리했습니다.
