+++ src/main/resources/mybatis/mapper/category-mapping.xml
@@ -37,4 +37,15 @@
 		values (#{main}, #{middle}, #{sub}, #{detail})
 	</insert>
 
+<!--사용자가 가장 많이 틀린 카테고리 -->
+	<select id="getWeakCategoryByCategorySeq" resultType="categorydto" parameterType="categorydto">
+		SELECT c.category_seq
+		FROM learn_record l
+					 INNER JOIN quiz q ON l.quiz_seq = q.quiz_seq
+					 INNER JOIN (SELECT category_seq, know_seq FROM knowledge) AS c ON c.know_seq = q.know_seq
+		WHERE is_right = 0
+			AND l.user_seq = ${userSeq}
+		GROUP BY category_seq
+		ORDER BY COUNT(*) DESC, category_seq LIMIT 0,1
+	</select>
 </mapper>
